move core logic from tools/pandoc_ast_diff.py to doc_build/ast_diff.py
